NotificationService Onboarding on Quarkus (#87)
* Chnages to send Just the Vulnerabily object and AnalyzerIdentity back to DT

* Configure Kafka Topic log retention to 24 hrs


* Configure Kafka Topic log retention to 24 hrs


* Notification Services Classes added to Project


* Notification Changes

* Added notification's generic config in application properties


* Unified the way we are reading the configs


* Changes to receive notification from notification topics


* WIP : Database intgration from Quarkus service


* Added Hibernate properties and dependency


* Maven dependency for Persistence annotations


* Removed non-quarkus dependency for Hibernate


* WIP

* wip- hibernate configuration changes

* query manager WIP

* Squashed commit of the following:

commit d72cbc5eb47d3367668789360f9514e8de830251
Author: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
Date:   Mon Nov 7 17:52:06 2022 +0000

    Update quarkus.platform.version to v2.13.4.Final (#75)


commit 968bc810fc64597ff8ef85df85f0302919ce7ae2
Author: Niklas <nscuro@protonmail.com>
Date:   Mon Nov 7 16:34:32 2022 +0100

    Fix Snyk client failing to encode PURLs properly (#74)

    * Fix Snyk client failing to encode PURLs properly

    The Snyk API requires *all* special characters to be URL encoded. However, the MicroProfile `@RestClient` would not encode `@` and `:`. It's not possible to customize the URL encoding, so for the time being we're stuck with writing out own client logic.

    Also re-added the handling of disabled analyzers. Analyzers that are not enabled will be part of the topology, and thus will not show up in the topology diagram.


    * Remove unused legacy analyzer code



commit d90bf9d9d564aefbfc05928fd118fbac5a2e36d8
Author: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
Date:   Sat Nov 5 10:32:05 2022 +0000

    Update returntocorp/semgrep-action digest to b470269 (#71)


commit 85042a4557323cbce707c7270952c7e80729b141
Author: mehab <meha.bhargava2@gmail.com>
Date:   Sat Nov 5 10:31:41 2022 +0000

    added kafka.bootstrap.servers=localhost:9092 property to enable kafka dev ui to work (#72)



commit 3442af19b55f5ab3697b4b2e408698963035063f
Author: Niklas <nscuro@protonmail.com>
Date:   Wed Nov 2 21:24:54 2022 +0100

    Rework Analyzer Topology (#66)

    * WIP state of reworking analyzer topology


    * Add some logging statements to demonstrate what's happening


    * Integrate simplified OSS Index analyzer into topology


    * Use existing model classes; Fix auth header for OSS Index


    * Remove unused classes; Cleanup serializers


    * Add simple topology test


    * Cleanup config properties


    * Provide a charset when calling `getBytes`


    * Remove dummy tests


    * Fix hardcoded maven path causing test failures


    * Tweak test script to use BOMs as input


    * Clarify Redpanda ports in docker-compose.yml


    * Document test script usage in README.md


    * Multiple improvements to topology


    * Remove redundant topic creation

    Kafka Streams takes care of this automatically.


    * Basic snyk analyzer implementation


    * Add rate limiting for Snyk


    * Add names to topology processors

    Not sure about the naming convention yet. Using short sentences is great for understanding the topology graph, but it may be a little too much.



commit 7649dbb95855b6b186343ec856da8118e3668172
Author: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
Date:   Mon Oct 31 10:57:42 2022 +0000

    Update registry.access.redhat.com/ubi8/ubi-minimal Docker tag to v8.6-994 (#67)


* Wip - Code to get rid of dependency on Alpine Framework


* Wip- Conversion of JDO Entity to JPA


* Wip- Conversion of JDO Entity to JPA


* Wip- Conversion of JDO Entity to JPA


* Wip- Conversion of JDO Entity to JPA


* Wip- Conversion of JDO Entity to JPA


* JDO to JPA WIP

* WIP2

* Create NotificationHibernateManagerTest.java

* Update VulnerabilityAnalysisEvent.java

* Wip- Conversion of JDO Entity to JPA


* Wip- Conversion of JDO Entity to JPA


* Wip- Conversion of JDO Entity to JPA


* Wip- Conversion of JDO Entity to JPA( Analysis, AnalysisComment, Bom, Vex, ViolationAnalysis, ViolationAnalysisComment, Vulnerability


* WIP

* fixed errors

* Wip- Conversion of JDO Entity to JPA


* Wip- Conversion of JDO Entity to JPA


* Wip- Conversion of JDO Entity to JPA


* Conversion of JDO Entity to JPA/Removal of alpine dependency


* Added annotation @Lob for columnDefinition="CLOB"


* Fixed few more annotations for Hibernate


* Added Hibernate Naming Strategy to be implicit


* Added Hibernate Naming Strategy to be implicit


* WIP2

* Fixed Hibernate Queries for reserved names


* Added Foreign key relationships


* Fix identifier quoting; Use testcontainers in test mode; Use API server schema in tests; Fix some field<->schema mismatches


* Refactored some unused persistent classes to non-persistent
Fixed the Schema Validation for Component


* Fixed few Schema Validation errors


* Fixed few Schema Validation errors


* Fixed Few Schema Validation errors


* Fixed all Schema Validation errors


* Update DefaultObjectGenerator.java

* Fixed the test for NotificationHibernateManagerTest


* query manager code refactoring

* Minor changes for config


* Fixed Schema validation error after NotificationManager changes


* Uncommneted call to load defaultNotificationPublisher


* Fixed the call to loadDefaultNotificationPublisher, can see NotificationPublisher object being returned from database


* Notification Service related changes


* Added @Transactional annotations ( Exception is still there though)


* Handling of streams for filtering Notification Rules


* Update database schema for tests

Includes data corrections regarding types of ID columns and foreign keys not matching up.


* Fix NotificationRule query


* Remove unused classes


* Remove redundant serde code


* Add missing `@Transactional` annotation for NotificationRouter


* Integrate notification consumer into existing topology

Having multiple separate streams topologies isn't really supported with Quarkus


* Remove initialization with default publishers

The application should not write to the database


* Remove more unused code


* Remove more unused code


* Temporary workaround for ID column type mismatches for Team model


* Add basic NotificationRouter test; Fix missing JSON-B provider


* Added Deserialization/Serialization ananotation for local Time in Notification


* Remove alpine as parent project; Cleanup POM


* Update NotificationRuleRepositoryTest to use native queries for setup


* 1. Added ConfigPropertyEntity
2. Added Repository for configProperty
3. Changes to send mail using Quarkus Mailer
4. Added Test for configPropertyRepository
5. Removed unused topic creation from docker file


* suppress warning for immutable field

+++ src/main/java/org/acme/common/IConfigProperty.java
@@ -1,9 +1,9 @@
 package org.acme.common;
 
 public interface IConfigProperty {
-    long getId();
+    int getId();
 
-    void setId(long var1);
+    void setId(int var1);
 
     String getGroupName();
